+++ packaging/chocolatey/tools/chocolateyInstall.ps1
@@ -1,0 +1,3 @@
+$ErrorActionPreference = 'Stop';
+$toolsDir = "$(Split-Path -parent $MyInvocation.MyCommand.Definition)"
+# Embedded package automatically shims netcheck.exe